refactor: move workflow based UI to workflows directory
The browser wallet will contain workflow based UI and objectives based UI
